soru5 modified.
