+++ src/soru5/Application.java
@@ -2,50 +2,48 @@ package soru5;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import soru2.Book;
 
-public class Application {
+public class Application2 {
 
 	public static void main(String[] args) {
 
-		ArrayList<Book> books = new ArrayList<Book>();
-		books.add(new Book("Java", "Steve", 2020, 260));
-		books.add(new Book("PHP", "Mahmud", 2015, 66));
-		books.add(new Book("Python", "Chuck", 2016, 510));
-		books.add(new Book("ML", "Sami", 2012, 70));
-		books.add(new Book("AI", "Ed", 2002, 480));
-		
-		books.add(new Book("DS", "Onur", 2020, 260));
-		books.add(new Book("CS", "Muhammed", 2015, 88));
-		books.add(new Book("SE", "Mahmud", 2016, 220));
-		books.add(new Book("CE", "Ahmed", 2012, 200));
-		books.add(new Book("CV", "Najeeb", 2002, 99));
-		
+		List<Book> books = new ArrayList<Book>();
 
-				System.out.println("Sayfa sayisi 100'den fazla olan kitaplar");
-				
-				Predicate<Book> pageCountMoreThan100 = (book) -> {
-					return (book.getPageCount() > 100 );
-				};
-				
-				Consumer<Book> printer = (book) -> {
-					System.out.println(book);
-				};
-				
-				//Map<String, String> booksMap = books.stream().forEach(collect(Collectors.toMap(books.ge, valueMapper)));
+		books.add(new Book("AAA", "Ali"  , 2020, 200));
+		books.add(new Book("BBB", "Bilal", 2019, 180));
+		books.add(new Book("CCC", "Can"  , 2018, 160));
+		books.add(new Book("DDD", "Deniz", 2017, 140));
+		books.add(new Book("EEE", "Emre" , 2016, 120));
+		books.add(new Book("FFF", "Faruk", 2015, 100));
+		books.add(new Book("GGG", "Gamze", 2014, 80));
+		books.add(new Book("HHH", "Hasan", 2013, 60));
+		books.add(new Book("III", "Ihsan", 2012, 40));
+		books.add(new Book("JJJ", "James", 2011, 20));
 
+		Stream<Book> stream = books.stream();
+		Consumer<? super Book> printer = (book) -> {System.out.println(book);};
+		System.out.println("Books without any filter:");
+		stream.forEach(printer);
 
-				List<Book> pageCountMoreThan100List = books.stream().filter(pageCountMoreThan100).collect(Collectors.toList());
+		Predicate<Book> greaterThan100 = (book) -> {
 
-				pageCountMoreThan100List.stream().forEach(printer);
-				
-				
-				
-				
+			if (book.getPageCount() > 100)
+				return true;
+
+			return false;
+		};
+
+		Set<Book> greaterThan100List = books.stream().filter(greaterThan100).collect(Collectors.toSet());
+		Consumer<Book> bookPrinter = (book) -> System.out.println(book);
+		System.out.println("\nBooks with more than 100 pages only:");
+		greaterThan100List.stream().forEach(bookPrinter);
 	}
 
 }
